feat: 542794 - Fix TS lint errors
