+++ designer/client/src/javascripts/preview/radio.test.js
@@ -220,6 +220,6 @@ describe('radio', () => {
 })
 
 /**
- * @import { ListElement } from '~/src/javascripts/preview/radio.js'
+ * @import { ListElement } from '@defra/forms-model'
  * @import { QuestionBaseModel } from '~/src/javascripts/preview/question.js'
  */
